+++ models/registry/work_item.go
@@ -22,9 +22,9 @@ type WorkItem struct {
 
 	// Read-only, from view.
 	GenericFileIdentifier string `json:"generic_file_identifier"`
-	ID                    int    `json:"id,omitempty"`
+	ID                    int64  `json:"id,omitempty"`
 	InstApprover          string `json:"inst_approver"`
-	InstitutionID         int    `json:"institution_id"`
+	InstitutionID         int64  `json:"institution_id"`
 	Name                  string `json:"name"`
 	NeedsAdminReview      bool   `json:"needs_admin_review"`
 	Node                  string `json:"node"`
@@ -98,7 +98,7 @@ type WorkItemForPharos struct {
 	ETag                  string     `json:"etag"`
 	GenericFileIdentifier string     `json:"generic_file_identifier"`
 	InstApprover          string     `json:"inst_approver"`
-	InstitutionID         int        `json:"institution_id"`
+	InstitutionID         int64      `json:"institution_id"`
 	Name                  string     `json:"name"`
 	NeedsAdminReview      bool       `json:"needs_admin_review"`
 	Node                  string     `json:"node"`
